companyservice/requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
mysqlclient==2.2.4
sqlparse==0.5.0
tzdata==2024.1
cryptography~=42.0.7
django-cors-headers~=4.1.0
drf-jwt~=1.13.3
pyjwt~=1.7.1
requests~=2.22.0
python-dotenv==1.0.1
pika==1.2.0